test: fold NVIDIA Model-Optimizer vit example into test_timm.py
Move the vit_base_patch16_224 case into test_timm.py alongside the existing
swin_tiny_patch4_window7_224 test (both are the models named in NVIDIA Model
Optimizer's onnx_ptq download_example_onnx.py), and drop the separate
test_modelopt.py file.
